Create skeleton Svelte app
