+++ .gitignore
@@ -1,4 +1,4 @@
-.svelte
+.svelte-kit
 .vercel
 build
