+++ package.json
@@ -1,32 +1,74 @@
 {
-  "name": "iconophor",
-  "version": "1.0.0",
-  "repository": "git@github.com:msfragala/iconophor.git",
-  "author": "Mitchell Fragala <ms.fragala@gmail.com>",
-  "private": true,
-  "scripts": {
-    "lint": "eslint .",
-    "prepare": "husky install"
-  },
-  "prettier": {
-    "trailingComma": "es5",
-    "tabWidth": 2,
-    "quoteProps": "consistent",
-    "singleQuote": true
-  },
-  "browserslist": [
-    "last 1 chrome version",
-    "last 1 firefox version",
-    "last 1 safari version",
-    "last 1 ios version",
-    "last 1 edge version",
-    "last 1 chromeandroid version"
-  ],
-  "dependencies": {
-    "hast-util-to-html": "^9.0.0",
-    "ms": "^2.1.3",
-    "node-fetch": "^3.3.2",
-    "svg-parser": "^2.0.4"
-  },
-  "devDependencies": {}
+	"name": "iconophor",
+	"version": "1.0.0",
+	"type": "module",
+	"repository": "git@github.com:msfragala/iconophor.git",
+	"author": "Mitchell Fragala <ms.fragala@gmail.com>",
+	"private": true,
+	"scripts": {
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview",
+		"lint": "eslint .",
+		"prepare": "husky install"
+	},
+	"prettier": {
+		"trailingComma": "es5",
+		"tabWidth": 2,
+		"useTabs": true,
+		"quoteProps": "consistent",
+		"singleQuote": true,
+		"plugins": [
+			"prettier-plugin-svelte"
+		],
+		"pluginSearchDirs": [
+			"."
+		],
+		"overrides": [
+			{
+				"files": "*.svelte",
+				"options": {
+					"parser": "svelte"
+				}
+			}
+		]
+	},
+	"browserslist": [
+		"last 1 chrome version",
+		"last 1 firefox version",
+		"last 1 safari version",
+		"last 1 ios version",
+		"last 1 edge version",
+		"last 1 chromeandroid version"
+	],
+	"dependencies": {
+		"hast-util-to-html": "^9.0.0",
+		"ms": "^2.1.3",
+		"node-fetch": "^3.3.2",
+		"svg-parser": "^2.0.4"
+	},
+	"devDependencies": {
+		"@sveltejs/adapter-auto": "^2.0.0",
+		"@sveltejs/kit": "^1.20.4",
+		"@typescript-eslint/eslint-plugin": "^6.0.0",
+		"@typescript-eslint/parser": "^6.7.5",
+		"autoprefixer": "^10.4.14",
+		"eslint": "^8.28.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-import-resolver-typescript": "^3.6.1",
+		"eslint-plugin-import": "^2.28.1",
+		"eslint-plugin-path-alias": "^1.0.0",
+		"eslint-plugin-svelte": "^2.34.0",
+		"husky": "^8.0.3",
+		"postcss": "^8.4.24",
+		"postcss-load-config": "^4.0.1",
+		"prettier": "^2.8.0",
+		"prettier-plugin-svelte": "^2.10.1",
+		"svelte": "^4.0.5",
+		"svelte-check": "^3.4.3",
+		"tailwindcss": "^3.3.2",
+		"tslib": "^2.4.1",
+		"typescript": "^5.2.2",
+		"vite": "^4.4.2"
+	}
 }
