+++ postcss.config.cjs
@@ -1,0 +1,9 @@
+const config = {
+	plugins: [
+		require('tailwindcss/nesting'),
+		require('tailwindcss'),
+		require('autoprefixer'),
+	],
+};
+
+module.exports = config;
