--- postcss.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-  plugins: {
-    'tailwindcss': {},
-    'postcss-pxtorem': {
-      propList: ['*'],
-      minPixelValue: 1,
-    },
-    'autoprefixer': {},
-  },
-};
