+++ src/routes/+page.svelte
@@ -1,0 +1,4 @@
+<h1>Welcome to SvelteKit</h1>
+<p>
+	Visit <a href="https://kit.svelte.dev">kit.svelte.dev</a> to read the documentation
+</p>
