+++ svelte.config.js
@@ -1,0 +1,15 @@
+import adapter from '@sveltejs/adapter-auto';
+import { vitePreprocess } from '@sveltejs/kit/vite';
+
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	preprocess: [vitePreprocess()],
+	kit: {
+		adapter: adapter(),
+		alias: {
+			'@': './src',
+		},
+	},
+};
+
+export default config;
